Fpga fixes (#82)
* fix: ppage truncation beyond 32 bit

* fix: shift make sure 64 bit instead of 32 bit

* fix: ldrsw size 64 instead of 32

+++ regression/src/client/fpga_aws.c
@@ -113,7 +113,7 @@ int initFPGAContext(FPGAContext *c) {
 
   // reset the DRAM page table
   uint8_t zero_page[4096] = {0};
-  for(int ppn = 0; ppn < (c->ppage_base_addr) / 4096; ++ppn){
+  for(uint64_t ppn = 0; ppn < (c->ppage_base_addr) / 4096; ++ppn){
     writeAXI(c, ppn << 12, zero_page, 4096);
   }
